requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
fastapi==0.115.5
idna==3.10
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
